# projects/support_ticketing_system/backend/backend_app/api/schemas/user.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""API request/response models for entity.user (from api.json schemas)."""
from __future__ import annotations

from datetime import datetime
from typing import Literal
from uuid import UUID

from pydantic import BaseModel, ConfigDict, EmailStr, Field


class UserResponse(BaseModel):
    """schema.user: User"""

    model_config = ConfigDict(from_attributes=True)
    id: UUID
    email: str
    full_name: str
    role: Literal['customer', 'agent', 'manager']
    created_at: datetime


class UserLoginRequest(BaseModel):
    """schema.user.login: Log in request"""

    model_config = ConfigDict(extra="forbid", str_strip_whitespace=True)
    email: EmailStr = Field(..., max_length=254, min_length=1)
    password: str = Field(..., max_length=128, min_length=1)


class UserRequestPasswordResetRequest(BaseModel):
    """schema.user.request_password_reset: Request password reset request"""

    model_config = ConfigDict(extra="forbid", str_strip_whitespace=True)
    email: EmailStr = Field(..., max_length=254, min_length=1)


class UserResetPasswordRequest(BaseModel):
    """schema.user.reset_password: Reset password request"""

    model_config = ConfigDict(extra="forbid", str_strip_whitespace=True)
    reset_token: str = Field(..., max_length=2048, min_length=1)
    new_password: str = Field(..., max_length=128, min_length=8)


class LoginSessionResponse(UserResponse):
    """Login response satisfying BOTH contracts: the graph's user schema (top-level fields) and the Frontend's Session {token, user}."""

    token: str
    token_type: str = "bearer"
    expires_in: int
    user: UserResponse