# projects/task_manager/backend/backend_app/api/routes/user.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""HTTP routes for service.user. Route handlers only delegate to the application service."""
from __future__ import annotations

from fastapi import APIRouter, Depends, Request, Response

from ...api_common import error_responses
from ...auth import authorize
from ...engine import Principal
from ..deps import get_services
from ...application.services import Services
from ..schemas.user import LoginSessionResponse, UserLoginRequest, UserRequestPasswordResetRequest, UserResetPasswordRequest


router = APIRouter(tags=["user"])


@router.post(
    "/auth/forgot-password",
    operation_id="api.user.request_password_reset",
    summary="Request password reset",
    description="Request password reset.\n\nOperation `operation.user.request_password_reset`. Access: public.",
    response_model=None,
    status_code=204,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=False, mutating=True, rate_limited=True),
    response_class=Response,
)
def api_user_request_password_reset(
    request: Request,
    body: UserRequestPasswordResetRequest,
    principal: None = Depends(authorize("operation.user.request_password_reset")),
    services: Services = Depends(get_services),
):
    services.user.request_password_reset(body, request.client.host if request.client else 'unknown')
    return Response(status_code=204)


@router.post(
    "/auth/login",
    operation_id="api.user.login",
    summary="Log in",
    description="Log in.\n\nOperation `operation.user.login`. Access: public.",
    response_model=LoginSessionResponse,
    status_code=200,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=False, mutating=True, rate_limited=True),
)
def api_user_login(
    request: Request,
    body: UserLoginRequest,
    principal: None = Depends(authorize("operation.user.login")),
    services: Services = Depends(get_services),
):
    return services.user.login(body, request.client.host if request.client else 'unknown')


@router.post(
    "/auth/logout",
    operation_id="api.user.logout",
    summary="Log out",
    description="Log out.\n\nOperation `operation.user.logout`. Access: authenticated.",
    response_model=None,
    status_code=204,
    responses=error_responses(authenticated=True, restricted=False, has_path_id=False, mutating=True, rate_limited=False),
    response_class=Response,
)
def api_user_logout(
    principal: Principal | None = Depends(authorize("operation.user.logout")),
    services: Services = Depends(get_services),
):
    services.user.logout(principal)
    return Response(status_code=204)


@router.post(
    "/auth/reset-password",
    operation_id="api.user.reset_password",
    summary="Reset password",
    description="Reset password.\n\nOperation `operation.user.reset_password`. Access: public.",
    response_model=None,
    status_code=204,
    responses=error_responses(authenticated=False, restricted=False, has_path_id=False, mutating=True, rate_limited=True),
    response_class=Response,
)
def api_user_reset_password(
    request: Request,
    body: UserResetPasswordRequest,
    principal: None = Depends(authorize("operation.user.reset_password")),
    services: Services = Depends(get_services),
):
    services.user.reset_password(body, request.client.host if request.client else 'unknown')
    return Response(status_code=204)